refactor: split adapters/react into hooks and sync subdirectories
Improve organization by separating adapter hooks by responsibility:
- adapters/react/hooks/ - Action hooks (useSettings, usePlayerControls, etc.)
- adapters/react/sync/ - Synchronization hooks (useEventSync, useLyricsSync, etc.)

Changes:
- Create hooks/ subdirectory for action/access hooks
- Create sync/ subdirectory for state synchronization hooks
- Add index.ts barrel export for cleaner imports
- Update all imports to use barrel export pattern

Action hooks moved to hooks/:
- usePlayerControls.ts
- useSettings.ts
- usePlaylists.ts
- useProviders.ts
- useEventBus.ts

Sync hooks moved to sync/:
- useEventSync.ts
- usePlayerSync.ts
- usePlayerControlSync.ts
- useLyricsSync.ts
- useArtworkSync.ts
- useProviderStatus.ts

Updated imports in 10 source files and 1 test helper.
All 155 tests passing.
