+++ client/src/adapters/react/hooks/useEventBus.ts
