+++ client/src/adapters/react/hooks/useSettings.ts
